requirements000.txt
altair==5.1.2
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
crispy-bootstrap4==2022.1
cryptography==41.0.4
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Django==4.2.4
django-admin-logs==1.0.2
django-crispy-forms==1.14.0
django-plotly-dash==2.2.2
dpd-components==0.1.0
dpd-static-support==0.0.5
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
Flask==3.0.0
fonttools==4.43.1
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
# networkx==3.1
nltk==3.8.1
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==1.3.5
parso==0.8.3
patsy==0.5.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
plotly==5.17.0
plotly-express==0.4.1
prompt-toolkit==3.0.39
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2018.7
pyvis==0.3.2
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.6
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
Sastrawi==1.0.1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.27.2
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
Werkzeug==3.0.1
whitenoise==6.5.0
wordcloud==1.9.2
zipp==3.17.0
